requirements.txt
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
